+++ app/src/main/java/com/api/EngineerCollabo/controllers/ChatApiController.java
@@ -2,33 +2,42 @@ package com.api.EngineerCollabo.controllers;
 
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import com.api.EngineerCollabo.entities.ChatMessage;
+import com.api.EngineerCollabo.entities.Message;
+import com.api.EngineerCollabo.entities.ResponseMessage;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import com.api.EngineerCollabo.services.ChatService;
-import com.api.EngineerCollabo.repositories.ChatRepository;
+import com.api.EngineerCollabo.services.MessageService;
+import com.api.EngineerCollabo.repositories.MessageRepository;
 
 @RestController
 @RequestMapping("/api/messages")
 public class ChatApiController {
 
     @Autowired
-    ChatRepository chatRepository;
+    MessageRepository messageRepository;
 
     @Autowired
-    ChatService chatService;
+    MessageService messageService;
 
     @PostMapping
-    public ResponseEntity<?> sendChatMessage(@RequestBody ChatMessage chatMessage) {
+    public ResponseEntity<?> sendMessage(@RequestBody Message message) {
         // メッセージを処理して保存するなどの処理を行う
-        chatRepository.save(chatMessage);
+        String text = message.getText();
+        String content = message.getContent();
+        Integer userId = message.getUserId();
+        Integer channelId = message.getChannelId();
+
+        if(userId !=null && channelId !=null){
+            messageService.saveMessage(text, content, userId, channelId);
+        }
         return ResponseEntity.ok().build();
+
     }
 
     @GetMapping
-    public List<ChatMessage> getChatMessages() {
+    public List<ResponseMessage> getAllMessages() {
         // チャットの履歴を取得する処理を行う
-        return chatService.getAllChatMessages();
+        return messageService.getAllMessages();
     }
 }
