--- app/src/main/java/com/api/EngineerCollabo/controllers/ChatController.java
@@ -1,15 +1,0 @@
-package com.api.EngineerCollabo.controllers;
-
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class ChatController {
-
-    @MessageMapping("/chat")
-    @SendTo("/topic/messages")
-    public String sendChatMessage(String chatMessage) {
-        return chatMessage;
-    }
-}
