--- app/src/main/java/com/api/EngineerCollabo/entities/ChatMessage.java
@@ -1,27 +1,0 @@
-package com.api.EngineerCollabo.entities;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Table;
-import jakarta.persistence.Column;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name ="chat_messages")
-public class ChatMessage {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name ="content")
-    private String content;
-
-    @Column(name = "sender")
-    private String sender;
-
-    public ChatMessage() {
-    }
-}
