+++ app/src/main/java/com/api/EngineerCollabo/entities/Message.java
@@ -57,4 +57,8 @@ public class Message {
     @JoinColumn(name = "channel_id", nullable = false, referencedColumnName = "id", insertable = false, updatable = false)
     private Channel channel;
 
+    public Message(){
+        
+    }
+
 }
