+++ app/src/main/java/com/api/EngineerCollabo/services/MessageService.java
@@ -3,6 +3,8 @@ package com.api.EngineerCollabo.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.ArrayList;
 import com.api.EngineerCollabo.entities.Channel;
 import com.api.EngineerCollabo.entities.Message;
 import com.api.EngineerCollabo.entities.ResponseMessage;
@@ -25,7 +27,12 @@ public class MessageService {
     @Autowired
     private MessageRepository messageRepository;
 
-    public Message createMessage(String text, String content, Integer userId, Integer channelId) {
+    @Autowired
+    public MessageService(MessageRepository messageRepository) {
+        this.messageRepository = messageRepository;
+    }
+
+    public void saveMessage(String text, String content, Integer userId, Integer channelId) {
         User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User not found"));
         Channel channel = channelRepository.findById(channelId)
                 .orElseThrow(() -> new EntityNotFoundException("Channel not found"));
@@ -35,18 +42,49 @@ public class MessageService {
         message.setUserId(user.getId());
         message.setChannelId(channel.getId());
 
-        return messageRepository.save(message);
+        messageRepository.save(message);
     }
 
-    public ResponseMessage changeResponseMessage(Message message) {
-        ResponseMessage responseMessage = new ResponseMessage();
-        responseMessage.setId(message.getId());
-        responseMessage.setText(message.getText());
-        responseMessage.setContent(message.getContent());
-        responseMessage.setUserId(message.getUser().getId());
-        responseMessage.setChannelId(message.getChannel().getId());
-        responseMessage.setCreatedAt(message.getCreatedAt());
-        responseMessage.setUpdatedAt(message.getUpdatedAt());
-        return responseMessage;
+    public List<ResponseMessage> getAllMessages(){
+        List<Message> messages = messageRepository.findAll();
+        List<ResponseMessage> responseMessages = new ArrayList<>();
+        for (Message message : messages){
+            ResponseMessage responseMessage = new ResponseMessage();
+            responseMessage.setId(message.getId());
+            responseMessage.setText(message.getText());
+            responseMessage.setContent(message.getContent());
+            responseMessage.setUserId(message.getUser().getId());
+            responseMessage.setChannelId(message.getChannel().getId());
+            responseMessage.setCreatedAt(message.getCreatedAt());
+            responseMessage.setUpdatedAt(message.getUpdatedAt());
+            responseMessages.add(responseMessage);
+        }
+        return responseMessages;
     }
+
+
+    // public Message createMessage(String text, String content, Integer userId, Integer channelId) {
+    //     User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User not found"));
+    //     Channel channel = channelRepository.findById(channelId)
+    //             .orElseThrow(() -> new EntityNotFoundException("Channel not found"));
+    //     Message message = new Message();
+    //     message.setText(text);
+    //     message.setContent(content);
+    //     message.setUserId(user.getId());
+    //     message.setChannelId(channel.getId());
+
+    //     return messageRepository.save(message);
+    // }
+
+    // public ResponseMessage changeResponseMessage(Message message) {
+    //     ResponseMessage responseMessage = new ResponseMessage();
+    //     responseMessage.setId(message.getId());
+    //     responseMessage.setText(message.getText());
+    //     responseMessage.setContent(message.getContent());
+    //     responseMessage.setUserId(message.getUser().getId());
+    //     responseMessage.setChannelId(message.getChannel().getId());
+    //     responseMessage.setCreatedAt(message.getCreatedAt());
+    //     responseMessage.setUpdatedAt(message.getUpdatedAt());
+    //     return responseMessage;
+    // }
 }
